Refactored file structure for sustainability
