--- src/classes/industry-item-tags.ts
@@ -1,19 +1,0 @@
-// The relationship between station industry and items
-
-import { ItemTag } from "./item-type";
-import { StationIndustry } from "./station";
-
-/*
-export type ItemTag =
-  | "food"
-  | "drink"
-  | "essential"
-  | "mining"
-  | "luxury"
-  | "components"
-  | "contraband"
-  | "materials"
-  | "medical"
-  | "ore",
-  | "goods";
-*/
