--- src/classes/screen-type.ts
@@ -1,8 +1,0 @@
-export type ScreenType =
-  | "StationInfoScreen"
-  | "StationTradeScreen"
-  | "StationMissionScreen"
-  | "StationMapScreen"
-  | "StationFuelScreen"
-  | "StationCrewScreen"
-  | "TravelScreen";
