+++ src/components/watermark.tsx
@@ -1,14 +1,9 @@
 export function Watermark() {
   return (
     <div className="watermark">
-      <img
-        src="./Retrograde_Logo.png"
-        alt="Retrograde Logo"
-        className="logo"
-        width={32}
-      />
+      <img src="./Retrograde_Logo.png" alt="Retrograde Logo" className="logo" width={32} />
       <div className="title">Retrograde</div>
-      <div className="version">v0.0.1</div>
+      <div className="version">v0.0.2</div>
     </div>
   );
 }
