+++ src/constants/index.ts
@@ -1,0 +1,12 @@
+export const REVOLUTIONS_IN_CYCLE = 400;
+
+export const defaultBuyPriceModifier = 1.0;
+export const defaultSellPriceModifier = 0.8;
+// Maximum price is 5x the baseline price
+export const maxPriceMultiplier = 5;
+
+export const MIN_STATION_DISTANCE = 12;
+export const MAX_STATION_DISTANCE = 32;
+
+// Travel speed is milliseconds per parsec
+export const TRAVEL_SPEED = 50;
