+++ src/logic/encounter.ts
