+++ src/logic/inventory.ts
@@ -1,0 +1,24 @@
+import { Inventory } from "../types/Inventory";
+import { ItemType } from "../types/ItemType";
+import { set } from "../utils/util";
+
+export function newInventory(): Inventory {
+  return {};
+}
+
+export function cloneInventory(inventory: Inventory): Inventory {
+  return { ...inventory };
+}
+
+export function getItemCount(inventory: Inventory, item: ItemType): number {
+  return inventory[item.name] ?? 0;
+}
+
+export function setItemCount(inventory: Inventory, item: ItemType, value: number): Inventory {
+  return set(inventory, { [item.name]: value });
+}
+
+export function addItemCount(inventory: Inventory, item: ItemType, value: number): Inventory {
+  const newCount = getItemCount(inventory, item) + value;
+  return setItemCount(inventory, item, newCount);
+}
