+++ src/types/ScreenType.ts
@@ -1,0 +1,8 @@
+export type ScreenType =
+  | "StationInfoScreen"
+  | "StationTradeScreen"
+  | "StationMissionScreen"
+  | "StationMapScreen"
+  | "StationFuelScreen"
+  | "StationCrewScreen"
+  | "TravelScreen";
